Handle case where model rank could be null
Here, the model rank could be null (before metrics kick in), and
it is in fact null, then tRPC would throw a type error complaining
that it tried to access properties of a null variable.

This causes no models to be shown on the homepage, at least not until
metrics kick in. (I was able to reproduce this myself)

In other words, for a brief moment (longest 1min), if a model is posted
then there is a possibility that no models would be shown on the
homepage.

I went with zero-ing the values since they probably would be zero in the
brief moments after they were posted anyway. And it doesn't matter too
much anyway because after metrics kick in, the 'actual' values will be
shown.